Use Map in favour of ImageMeta
The metadata can be of variable length. If we hardcode this, we would need to
update the code when unsupported values are encountered.

+++ src/main/java/com/afrozaar/wordpress/wpapi/v2/model/Media.java
@@ -43,7 +43,7 @@ public class Media {
     @JsonProperty("media_type")
     private String mediaType;
     @JsonProperty("post")
-    private Integer post;
+    private Long post;
     @JsonProperty("source_url")
     private String sourceUrl;
 
@@ -95,11 +95,11 @@ public class Media {
         this.mediaType = mediaType;
     }
 
-    public Integer getPost() {
+    public Long getPost() {
         return post;
     }
 
-    public void setPost(Integer post) {
+    public void setPost(Long post) {
         this.post = post;
     }
